Consolidate convertProfileToShellLaunchConfig into instance service
Part of #136062
